allow for multiple onloads
now exports function that receives root node (default: document.body)
and returns what was previously exported (the `onload` function).

this allows mutation observations to happen when the export is explicitly
called, rather than when the module is required. because of this, there
is now also a `.end()` function attached to the `onload` function
returned which disconnects the mutation observer.